requirements.txt
pandas
openai
python-dotenv
Pillow
